handles other sets
