+++ src/components/XenaGeneSetApp.js
@@ -245,77 +245,6 @@ export default class XenaGeneSetApp extends PureComponent {
 
     };
 
-    addGeneSet = (selectedPathway) => {
-      const selectedPathwaySet = update(this.state.pathwaySet, {
-        pathways: { $unshift: [{
-          goid: '',
-          golabel: selectedPathway,
-          gene: [],
-        }] }
-      });
-      AppStorageHandler.storePathways(selectedPathwaySet.pathways);
-      this.setState({
-        pathwaySet: selectedPathwaySet,
-        pathways: selectedPathwaySet.pathways,
-        fetch: true,
-        currentLoadState: LOAD_STATE.LOADING,
-      });
-    };
-
-    addGene = (selectedPathway, selectedGene) => {
-      // get pathway to filter
-      let pathwayIndex = this.state.pathwaySet.pathways.findIndex(p => selectedPathway.golabel === p.golabel);
-      let newSelectedPathway = update(this.state.pathwaySet.pathways[pathwayIndex],{
-        gene: { $unshift: [selectedGene]}
-      });
-      let selectedPathwaySet = update(this.state.pathwaySet, {
-        pathways:  { [pathwayIndex]: {$set:newSelectedPathway }}
-      });
-
-      AppStorageHandler.storePathways(selectedPathwaySet.pathways);
-      this.setState({
-        pathwaySet: selectedPathwaySet,
-        pathways: selectedPathwaySet.pathways,
-        selectedPathway: newSelectedPathway,
-        fetch: true,
-        currentLoadState: LOAD_STATE.LOADING,
-      });
-    };
-
-    removeGene = (selectedPathway, selectedGene) => {
-      const pathwayIndex = this.state.pathwaySet.pathways.findIndex(p => selectedPathway.golabel === p.golabel);
-      const geneIndex = this.state.pathwaySet.pathways[pathwayIndex].gene.findIndex(g => g !== selectedGene);
-      let newSelectedPathway = update(this.state.pathwaySet.pathways[pathwayIndex],{
-        gene: { $splice: [[geneIndex,1]]}
-      });
-      let selectedPathwaySet = update(this.state.pathwaySet, {
-        pathways:  { [pathwayIndex]: {$set:newSelectedPathway }}
-      });
-      AppStorageHandler.storePathways(selectedPathwaySet.pathways);
-      this.setState({
-        pathwaySet: selectedPathwaySet,
-        pathways: selectedPathwaySet.pathways,
-        selectedPathway: newSelectedPathway,
-        fetch: true,
-        currentLoadState: LOAD_STATE.LOADING,
-      });
-    };
-
-    removeGeneSet = (selectedPathway) => {
-      const pathwayIndex  = this.state.pathwaySet.pathways.findIndex(p => selectedPathway.golabel === p.golabel);
-      const selectedPathwaySet = update(this.state.pathwaySet, {
-        pathways: { $splice: [[pathwayIndex,1]] }
-      });
-      AppStorageHandler.storePathways(selectedPathwaySet.pathways);
-      this.setState({
-        pathwaySet: selectedPathwaySet,
-        pathways: selectedPathwaySet.pathways,
-        selectedPathway: undefined,
-        fetch: true,
-        currentLoadState: LOAD_STATE.LOADING,
-      });
-    };
-
     getActiveApp() {
       return this.state.pathwaySet;
     }
